WIP refactored in response to PR #202
