--- apps/coscrad-frontend/src/components/AllResources/AllResources.module.scss
@@ -1,3 +1,0 @@
-.resourcetypes {
-    border: 1px solid black;
-}
