+++ apps/coscrad-frontend/src/components/presenters/ResourceInfoPresenter/ResourceInfoPresenter.css
@@ -1,4 +1,3 @@
 .resource-meta {
     padding-left: 20px;
-    border: 1px solid black;
 }
