+++ apps/coscrad-frontend/src/configurable-front-matter/data/configurable-front-matter.json
@@ -1,7 +1,7 @@
 {
-    "siteTitle": "Aaron's Web of Knowledge",
-    "subTitle": "A Coscrad Testing Playground",
-    "about": "This is my test instance of the COSCRAD web of knowledge UI!",
-    "siteDescription": "Aaron comes here to ensure that all is well.",
+    "siteTitle": "Cultural Research Web of Knowledge",
+    "subTitle": "A Testing Instance of the COSCRAD Platform",
+    "about": "This is a test instance of the COSCRAD web of knowledge UI.  COSCRAD stands for Collective of Open Source Cultural Research and Development and is a team of creatives, software engineers and developers who are building this Web of Knowledge Platform to support different forms of community-based cultural research and archiving.  For more information please contact coscrad@gmail.com",
+    "siteDescription": "The COSCRAD Cultural Research Platform is a tool for community-based research that enables flexible connections between distinct types of content and media.",
     "copyrightHolder": "COSCRAD"
 }
