+++ apps/coscrad-frontend/src/configurable-front-matter/data/sample-front-matter.json
